Install jest for testing (and more.  See below.)
This change does a few things:
 - Install jest so I can start writing tests.
 - Use the module version of matter-js, poly-decomp, and p5 instead of
   using the client scripts.  This was done so the main plinko.js file
   could require them when testing on the server side.
 - Use browserify to compile backend javascripts for use in the browser.
 - Add watchify and nodeman to watch for file changes and restart the
   node server on those changes.

NOTE: The reason I'm taking a snapshot here, without an actual test that
tests the plinko logic, is because I'm having a hard time accessing
individual methods in the main 'App' object.  I assume this is because
the wy I wrote the main plinko script may be unconventional and
therefore harder to use in a testing context.