Remove logic to check time of input versus game start
Now that the app is using websockets and irc, and pretty much getting
input in real-time, I've removed logic that was leftover from when the
game polled for comments and had to process several comments at once,
which were not in real-time.

Time is still stored on the chip object for debugging purposes.  If we
need to know a player's most recent played chip, it can be checked on
the most recent chip.