Fix bug where score occassionally isn't tallied
Previously, sometimes a chip would score, but the score was never
tallied.  This was because the score event was triggered when a
collision between a chip and another body happened.  However, the
'collosionStart' event triggered by matter-js can have more than one
pair of bodies colliding.  In the event that two collisions happened, at
the same time, only the first collision was checked to see if it score.
This resulted in other collisions at the same time not being checked to
see if they scored.

This change iterates through all collisions from the event and checks
for the score, instead of just the first collision.  This should prevent
any collisinos with score sensors from being ignored.